Lab10HW

+++ Lab4HW/Lab4HW/Lab4HW.c
@@ -15,7 +15,7 @@
 //      res - общий заработок
 //--------------------------------------------------------
 
-double result(int b, int c, double a, double x )
+double result(int b, int c, int a, int x )
 {
     double res;
     res = a + (b - c) * x;
@@ -27,22 +27,22 @@ void main()
 {
 	setlocale(LC_CTYPE, "RUS");
 
-    double a, x, res;
+    int a, x, res;
 
     int b = 133;
     int c = 75;
 
     //Ввод данных
     printf("Введите заработок на первых 75 газетах: ");
-    scanf("%lf", &a);
+    scanf("%d", &a);
     printf("Введите заработок на каждой из остальных газет: ");
-    scanf("%lf", &x);
+    scanf("%d", &x);
 
     //вычисления
     res = result(b,c,a,x);
 
     //Вывод результата
-    printf("Общий заработок мальчика за продажу %d газет: %.2f руб.\n", b, res);
+    printf("Общий заработок мальчика за продажу %d газет: %d руб.\n", b, res);
 
     system("pause");
 }
